requirements.txt
pageindex
anthropic
pymupdf
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
PyJWT
bcrypt
